+++ OrbitGl/LiveFunctionDataView.cpp
@@ -185,11 +185,11 @@ void LiveFunctionsDataView::OnContextMenu( const std::wstring & a_Action, int a_
 //-----------------------------------------------------------------------------
 void LiveFunctionsDataView::OnFilter( const std::wstring & a_Filter )
 {
-    std::vector<int> indices;
+    std::vector<uint32_t> indices;
 
     std::vector< std::wstring > tokens = Tokenize( ToLower( a_Filter ) );
 
-    for( size_t i = 0; i < m_Functions.size(); ++i )
+    for( uint32_t i = 0; i < (uint32_t)m_Functions.size(); ++i )
     {
         const Function* function = m_Functions[i];
         if( function )
@@ -225,7 +225,7 @@ void LiveFunctionsDataView::OnFilter( const std::wstring & a_Filter )
 
     // Filter drawn textboxes
     Capture::GVisibleFunctionsMap.clear();
-    for( size_t i = 0; i < m_Indices.size(); ++i )
+    for( uint32_t i = 0; i < (uint32_t)m_Indices.size(); ++i )
     {
         Function & func = GetFunction(i);
         Capture::GVisibleFunctionsMap[func.GetVirtualAddress()] = &func;
@@ -239,7 +239,7 @@ void LiveFunctionsDataView::OnDataChanged()
 {
     size_t numFunctions = Capture::GFunctionCountMap.size();
     m_Indices.resize(numFunctions);
-    for (size_t i = 0; i < numFunctions; ++i)
+    for (uint32_t i = 0; i < numFunctions; ++i)
     {
         m_Indices[i] = i;
     }
